KY-36276

+++ solr/core/src/java/org/apache/solr/handler/loader/JsonLoader.java
@@ -575,8 +575,7 @@ public class JsonLoader extends ContentStreamLoader {
         case JSONParser.LONG:
           return parser.getLong();
         case JSONParser.NUMBER:
-//          return parser.getDouble();
-          return parser.getFloat();          
+          return parser.getDouble();
         case JSONParser.BIGNUMBER:
           return parser.getNumberChars().toString();
         case JSONParser.BOOLEAN:
